[screen] launch editor (GUI) from the command window
Added back the launching of gvim when the profile loads.

Problem is with i3wm, when running the terminal full-screen is not passing
correct window size to the launched application (or something like that).

No real solution yet, but mitigating the issue by temporally turning off the
full-screen mode, or moving the focus to different screen while the profile
loads.