[bash] bugfix in identifying git branch
When in a detached head `git status` return a message in the form of "HEAD
detached at nnnnnn", meaning current filtration return the word "at" (3rd
word).

Changed it to always return last word in the line, when in a proper ref it would
be the ref name, and in a detached head it will return the commit hash.